refactor: spring rest docs 관련 클래스 패키지 수정
- docs 패키지 생성
- EnumType 및 TestController 클래스를 docs로 이동
